TaskMate/upload-ai-feature/requirements.txt
flask
flask-cors
werkzeug
python-dotenv
requests
gunicorn
replicate